add mentor's comments
