+++ src/Root.tsx
@@ -11,9 +11,9 @@ export const Root = () => {
       <Route path="/" element={<App />}>
         <Route index element={<HomePage />} />
         <Route path="people">
-          <Route path=":slagId?" element={<PeoplePage />} />
+          <Route path=":slug?" element={<PeoplePage />} />
         </Route>
-        <Route path="home" element={<Navigate to={'/'} replace />} />
+        <Route path="home" element={<Navigate to="/" replace />} />
         <Route path="*" element={<NotFoundPage />} />
       </Route>
     </Routes>
